requirements.txt
azure-mgmt-compute
azure-mgmt-network
azure-mgmt-resource
azure-identity